requirements.txt
Flask==3.0.3
matplotlib==3.7.2
numpy==1.24.4
Pillow==10.4.0
Requests==2.32.3
torch==2.4.1
torchvision==0.20.0
tqdm==4.67.1
Werkzeug==3.0.3
